+++ HISTORY.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Fix `isFinished(req)` when data buffered
+
 2.2.0 / 2014-12-22
 ==================